requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data science
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Vector similarity search
faiss-cpu>=1.7.4
# For GPU support (choose the version matching your CUDA):
# pip install faiss-gpu

# LLM APIs (needed for Pipelines A & B)
openai>=1.0.0          # also used for Qwen via DashScope compatible endpoint
google-genai>=0.3.0    # Gemini API

# Visualization & evaluation
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
psutil>=5.9.0
httpx>=0.24.0
requests>=2.28.0
scipy>=1.7.0

# Transformer-based utilities (tokenisation helpers)
transformers
sentencepiece
